organize stacktrace code
