+++ backend/stacktraces/enhancer.go
@@ -1,4 +1,4 @@
-package errors
+package stacktraces
 
 import (
 	"context"
