+++ backend/stacktraces/enhancer_test.go
@@ -1,4 +1,4 @@
-package errors
+package stacktraces
 
 import (
 	"context"
