+++ backend/stacktraces/test-files/lodash.js
