+++ backend/stacktraces/test-files/vendors.js
